pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "clinical-contract"
version = "0.0.1"
description = "Valide des data contracts YAML cliniques contre des fichiers Parquet"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
license-files = ["LICENSE"]
keywords = ["data-contract", "clinical", "parquet", "validation", "quality"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

# La commande terminal
[project.scripts]
clinical-contract = "clinical_contract.cli:main"

[project.optional-dependencies]
duckdb = ["duckdb>=0.10", "pyarrow>=14.0"]
polars = ["duckdb>=0.10", "polars>=0.20", "pyarrow>=14.0"]
pyarrow = ["duckdb>=0.10", "pyarrow>=14.0"]
all = ["duckdb>=0.10", "polars>=0.20", "pyarrow>=14.0"]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "duckdb>=0.10",
    "polars>=0.20",
    "pyarrow>=14.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/clinical_contract"]
include = ["src/clinical_contract/py.typed"]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E9", "F63", "F7", "F82"]


[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
